requirements.txt
fastapi
uvicorn
python-multipart
torch
monai
antspyx
nibabel
numpy
pyvista
plotly
scikit-image
scipy
nilearn
